backend/BitflowConnectTime-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart